avoid compiler warning (implicit cast from time_t to long)
SVN-Revision: 25684

+++ v3p/dcmtk/dcmdata/libsrc/dcuid.cxx
@@ -1044,7 +1044,7 @@ char* dcmGenerateUniqueIdentifier(char* uid, const char* prefix)
     sprintf(buf, ".%lu", forcePositive(getpid()));
     addUIDComponent(uid, buf);
 
-    sprintf(buf, ".%lu", forcePositive(time(NULL)));
+    sprintf(buf, ".%lu", forcePositive(long(time(NULL))));
     addUIDComponent(uid, buf);
 
     sprintf(buf, ".%u", counter);
